Refactor chat render responsibilities
