+++ lib/widgets/mixin_image.dart
@@ -1,76 +1,12 @@
-import 'dart:async';
 import 'dart:io';
-import 'dart:ui' as ui;
+import 'dart:typed_data';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import '../utils/extension/extension.dart';
-import '../utils/image.dart';
-import '../utils/proxy.dart';
+import 'media_image_pipeline.dart';
 
 typedef PlaceholderWidgetBuilder = Widget Function();
-
-final _normalizedGifFiles = <String>{};
-
-@immutable
-class ProxyNetworkImage extends ImageProvider<ProxyNetworkImage> {
-  const ProxyNetworkImage(this.url, {this.scale = 1.0, this.proxyConfig});
-
-  final String url;
-  final double scale;
-  final ProxyConfig? proxyConfig;
-
-  @override
-  Future<ProxyNetworkImage> obtainKey(ImageConfiguration configuration) =>
-      SynchronousFuture<ProxyNetworkImage>(this);
-
-  @override
-  ImageStreamCompleter loadImage(
-    ProxyNetworkImage key,
-    ImageDecoderCallback decode,
-  ) => _load(key, (buffer) => decode(buffer));
-
-  ImageStreamCompleter _load(
-    ProxyNetworkImage key,
-    Future<ui.Codec> Function(ui.ImmutableBuffer buffer) decode,
-  ) => MultiFrameImageStreamCompleter(
-    codec: _loadAsync(key, decode),
-    scale: key.scale,
-    debugLabel: key.url,
-    informationCollector: () => <DiagnosticsNode>[
-      DiagnosticsProperty<ImageProvider>('Image provider', this),
-      DiagnosticsProperty<ProxyNetworkImage>('Image key', key),
-    ],
-  );
-
-  Future<ui.Codec> _loadAsync(
-    ProxyNetworkImage key,
-    Future<ui.Codec> Function(ui.ImmutableBuffer buffer) decode,
-  ) async {
-    try {
-      final bytes = normalizeGifBytesIfNeeded(
-        await downloadImageBytes(key.url, proxyConfig: key.proxyConfig),
-      );
-      return decode(await ui.ImmutableBuffer.fromUint8List(bytes));
-    } catch (_) {
-      scheduleMicrotask(() {
-        PaintingBinding.instance.imageCache.evict(key);
-      });
-      rethrow;
-    }
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      other is ProxyNetworkImage &&
-      other.url == url &&
-      other.scale == scale &&
-      other.proxyConfig == proxyConfig;
-
-  @override
-  int get hashCode => Object.hash(url, scale, proxyConfig);
-}
 
 class MixinImage extends StatelessWidget {
   const MixinImage({
@@ -157,7 +93,7 @@ class MixinImage extends StatelessWidget {
           errorBuilder?.call(context, error, stackTrace) ?? fallback(),
     );
 
-    return _LazyNormalizedGif(
+    return NormalizedGifImageGate(
       image: resolvedImage,
       placeholder: fallback,
       childBuilder: imageView,
@@ -175,80 +111,5 @@ class MixinImage extends StatelessWidget {
       );
     }
     return image;
-  }
-}
-
-class _LazyNormalizedGif extends StatefulWidget {
-  const _LazyNormalizedGif({
-    required this.image,
-    required this.placeholder,
-    required this.childBuilder,
-  });
-
-  final ImageProvider image;
-  final Widget Function() placeholder;
-  final Widget Function() childBuilder;
-
-  @override
-  State<_LazyNormalizedGif> createState() => _LazyNormalizedGifState();
-}
-
-class _LazyNormalizedGifState extends State<_LazyNormalizedGif> {
-  Future<void>? _future;
-
-  @override
-  void initState() {
-    super.initState();
-    _start();
-  }
-
-  @override
-  void didUpdateWidget(covariant _LazyNormalizedGif oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.image != widget.image) {
-      _start();
-    }
-  }
-
-  void _start() {
-    final image = widget.image;
-    if (image is! FileImage ||
-        !image.file.path.toLowerCase().endsWith('.gif')) {
-      _future = null;
-      return;
-    }
-
-    final path = image.file.absolute.path;
-    if (_normalizedGifFiles.contains(path)) {
-      _future = null;
-      return;
-    }
-
-    _future = _normalize(image.file);
-  }
-
-  Future<void> _normalize(File file) async {
-    try {
-      await normalizeGifFileIfNeeded(file, ImageType.gif.mimeType);
-      await FileImage(file).evict();
-    } finally {
-      _normalizedGifFiles.add(file.absolute.path);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final future = _future;
-    if (future == null) return widget.childBuilder();
-
-    return FutureBuilder<void>(
-      future: future,
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.done) {
-          return widget.childBuilder();
-        }
-        return widget.placeholder();
-      },
-    );
   }
 }
