+++ test/widgets/message/message_context_test.dart
@@ -1,8 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_app/constants/brightness_theme_data.dart';
 import 'package:flutter_app/db/mixin_database.dart';
+import 'package:flutter_app/ui/provider/database_provider.dart';
+import 'package:flutter_app/ui/provider/mention_cache_provider.dart';
+import 'package:flutter_app/ui/provider/setting_provider.dart';
+import 'package:flutter_app/widgets/brightness_observer.dart';
+import 'package:flutter_app/widgets/message/item/quote_message.dart';
 import 'package:flutter_app/widgets/message/message.dart';
+import 'package:flutter_app/widgets/message/message_action_policy.dart';
+import 'package:flutter_app/widgets/message/message_bubble.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:mixin_bot_sdk_dart/mixin_bot_sdk_dart.dart';
 
 void main() {
@@ -20,6 +29,29 @@ void main() {
 
     expect(rows.top.single.sameUserNext, isTrue);
     expect(rows.bottom.first.sameUserPrev, isTrue);
+  });
+
+  test('MessageActionPolicy keeps chat-only actions out of pinned views', () {
+    final message = testMessage('1');
+
+    final chatPolicy = MessageActionPolicy(
+      message: message,
+      isTranscriptPage: false,
+      isPinnedPage: false,
+      role: Object(),
+    );
+    expect(chatPolicy.canReply, isTrue);
+    expect(chatPolicy.canPin, isTrue);
+    expect(chatPolicy.canDelete, isTrue);
+
+    final pinnedPolicy = MessageActionPolicy(
+      message: message,
+      isTranscriptPage: false,
+      isPinnedPage: true,
+      role: Object(),
+    );
+    expect(pinnedPolicy.canReply, isFalse);
+    expect(pinnedPolicy.canDelete, isFalse);
   });
 
   testWidgets('MessageContext updates const children through inherited state', (
@@ -49,6 +81,51 @@ void main() {
 
     expect(find.text('2'), findsOneWidget);
   });
+
+  testWidgets('MessageBubble stretches short quote to message width', (
+    tester,
+  ) async {
+    const bodyKey = ValueKey('body');
+    final message = testMessage(
+      '1',
+      quoteId: 'quoted',
+      quoteContent: _quoteContent,
+    );
+
+    await tester.pumpWidget(
+      ProviderScope(
+        overrides: [
+          databaseProvider.overrideWith((ref) => DatabaseOpener()),
+          mentionCacheProvider.overrideWithValue(MentionCache(null)),
+          settingProvider.overrideWith((ref) => SettingChangeNotifier()),
+        ],
+        child: MaterialApp(
+          home: BrightnessData(
+            value: 0,
+            brightnessThemeData: lightBrightnessThemeData,
+            child: Align(
+              alignment: Alignment.topLeft,
+              child: MessageContext.fromMessageItem(
+                message: message,
+                child: const SizedBox(
+                  width: 400,
+                  child: MessageBubble(
+                    child: SizedBox(key: bodyKey, width: 260, height: 20),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+    await tester.pump();
+
+    final quoteWidth = tester.getSize(find.byType(QuoteMessage)).width;
+    final bodyWidth = tester.getSize(find.byKey(bodyKey)).width;
+
+    expect(quoteWidth, greaterThanOrEqualTo(bodyWidth));
+  });
 }
 
 class _MessageIdText extends HookWidget {
@@ -60,15 +137,29 @@ class _MessageIdText extends HookWidget {
   );
 }
 
-MessageItem testMessage(String id, {String userId = 'user'}) => MessageItem(
+MessageItem testMessage(
+  String id, {
+  String userId = 'user',
+  String? quoteId,
+  String? quoteContent,
+}) => MessageItem(
   messageId: id,
   conversationId: 'conversation',
   type: 'PLAIN_TEXT',
+  content: 'body',
   createdAt: DateTime(2026),
   status: MessageStatus.read,
   userId: userId,
   userIdentityNumber: '0',
   isVerified: false,
   sharedUserIsVerified: false,
+  quoteId: quoteId,
+  quoteContent: quoteContent,
   pinned: false,
 );
+
+const _quoteContent =
+    '{"message_id":"quoted","conversation_id":"conversation",'
+    '"user_id":"00000000-0000-4000-8000-000000000001",'
+    '"user_full_name":"KC","user_identity_number":"1","type":"PLAIN_TEXT",'
+    '"content":"q","createdAt":1767225600000,"status":"READ"}';
